+++ apps/doc/src/app/components/input/input-text/input.component.ts
@@ -23,6 +23,13 @@ export class InputComponent {
   public status: ZuiInputStatus = 'default';
   public statuses: ZuiInputStatus[] = ['default', 'success', 'warning', 'danger'];
 
+  public forceShowClearButtonVariants: ReadonlyArray<boolean | null> = [
+    null,
+    false,
+    true
+  ];
+  public forceShowClearButton = this.forceShowClearButtonVariants[0];
+
   public readonly zyfraInputBasicExample: TuiDocExample = {
     TypeScript: import('!!raw-loader!./examples/input-basic-example/input-basic-example.component.ts'),
     HTML: import('!!raw-loader!./examples/input-basic-example/input-basic-example.component.html'),