+++ libs/next/src/lib/components/chips/chips.module.ts
@@ -1,10 +1,21 @@
 import { CommonModule } from '@angular/common';
 import { NgModule } from '@angular/core';
 import { ZuiChipsComponent } from './chips.component';
+import { ZuiCallFuncModule } from '@digital-plant/zyfra-helpers';
+import { ZuiHintModule } from '../../directives';
+
 @NgModule({
-  imports: [CommonModule],
-  declarations: [ZuiChipsComponent],
-  exports: [ZuiChipsComponent],
+  imports: [
+    CommonModule,
+    ZuiCallFuncModule,
+    ZuiHintModule
+  ],
+  declarations: [
+    ZuiChipsComponent
+  ],
+  exports: [
+    ZuiChipsComponent
+  ],
 })
 export class ZuiChipsModule {}