+++ libs/next/src/lib/components/dropdowns/select/select.component.ts
@@ -14,8 +14,8 @@ import {
 import { ZuiDestroyService } from '@digital-plant/zyfra-helpers';
 import { FormControl, NgControl } from '@angular/forms';
 import { PolymorphContent } from '../../../directives';
-import { ZUI_SELECT_OPTIONS, ZuiSelectOptions } from './select.options';
-import { ZuiContextWithImplicit, ZuiFocusableElementAccessor, ZuiNativeFocusableElement } from '../../../types';
+import { ZUI_SELECT_OPTIONS, ZuiSelectOptions, ZuiSelectValueContext } from './select.options';
+import { ZuiFocusableElementAccessor, ZuiNativeFocusableElement } from '../../../types';
 import { ZuiInputSize } from '../../input';
 import { AbstractZuiControl } from '../../../abstract/control';
 import { zuiIsNativeFocused, zuiIsTextOverflow$ } from '../../../util';
@@ -97,11 +97,11 @@ implements ZuiFocusableElementAccessor
 
   @Input()
   @zuiDefaultProp()
-  emptyContent: string = this.options.emptyContent;
+  emptyContent: PolymorphContent = this.options.emptyContent;
 
   @Input()
   @zuiDefaultProp()
-  nullContent: string = this.options.nullContent;
+  nullContent: PolymorphContent = this.options.nullContent;
 
   readonly zuiIsTextOverflow$ = zuiIsTextOverflow$;
 
@@ -120,7 +120,7 @@ implements ZuiFocusableElementAccessor
 
   @Input()
   @zuiDefaultProp()
-  valueTemplate: PolymorphContent<ZuiContextWithImplicit<T> & {nullContent: string}> = this.options.valueContent;
+  valueTemplate: PolymorphContent<ZuiSelectValueContext<T>> = this.options.valueContent;
 
   @Input()
   @zuiDefaultProp()
@@ -160,7 +160,7 @@ implements ZuiFocusableElementAccessor
           this.dropdownHostElement?.reCalculatePositions(1000/60);
         }),
         debounceTime(0),
-        tap(() => this.safeOpenModal())
+        // tap(() => this.safeOpenModal())
       )
     }),
   );
@@ -206,7 +206,7 @@ implements ZuiFocusableElementAccessor
       tap((value) => {
         if (value) {
           value = this.items$.value?.find(
-            i => this.identityMatcher(value, i),
+            i => value && i && this.identityMatcher(value, i),
           );
         }
         this.select(value);
@@ -234,34 +234,34 @@ implements ZuiFocusableElementAccessor
 
   public select(item: T): void {
     this.updateValue(item);
-    this.requiredInputControl.setValue(this.stringify(item));
+    this.requiredInputControl.setValue(item && this.stringify(item));
     this.open = false;
   }
 
   public safeOpenModal(): void {
     const inputElement = this.focusableElement.nativeElement;
-    if (this.stop$.value) return
-    if (
+    // if (this.stop$.value) return
+    const open = (
       !this.open &&
       this.interactive &&
       inputElement &&
       zuiIsNativeFocused(inputElement)
-    ) {
-      this.open = true;
-      this.changeDetectorRef.markForCheck();
-    }
+    )
+    this.open = open;
+    console.log('#mz safeOpenModal open', {
+      open
+    })
+    this.changeDetectorRef.markForCheck();
   }
 
   // TODO remove after finish activezone to dropdown component
   public safeStopPropagation(value: string, $event: Event): void {
+    console.log('#mz safeStopPropagation', {
+      value,
+      $event
+    });
+    this.open = false;
+    this.changeDetectorRef.markForCheck();
     if (!value) $event.stopImmediatePropagation();
-    this.stop$.next(true);
-    timer(0).pipe(
-      tap(() => {
-        this.focusableElement.nativeElement.blur();
-        this.stop$.next(false)
-      }),
-      takeUntil(this.destroy$)
-    ).subscribe();
   }
 }